Added data to flyway scripts to create cohorts, trainers and trainees
Updated the Keycloak configuration to include these new users and added roles in keycloak to represent the cohorts.

The keycloak trainee and trainer users are associated with cohorts in keycloak and this is reflected in the flyway scripts to set up corresponding data on our Postgres database. This should be changed to be an administrative process that can be carried out through the portal at a later point in time.